Appended documentation to Gender
We advise against sending 'c' for 'child' or 'i' for 'infant', but this
can be sent in the XML, is interpreted appropriately and some users of
the interface send data in this way. Better is sending the Child or Baby
nodes for a Child or a Baby. If the child or baby is a male or female,
an 'm' or 'f' can then be sent. If the gender is unknown, then the
Gender node should not be sent.