New Dossier nodes and attributes
Added a Department node, which stores the name of the department that is
to receive the order. This is a free text field.
Added an attribute IsDeleted. If the value is true the dossier no longer
exists in the database. External systems that import OLTDossiers can use
this attribute as a trigger to delete the dossier in their systems.