DaysBeforeDepartureFrom > DaysBeforeDepartureTo: Documentation
This number is mapped to a DateBeforeDepartureTo/From date, i.e. current
date minus "to" number of days, so the value of DaysBeforeDepartureFrom
must be larger than the value of DaysBeforeDepartureTo as the FromDate
mus be smaller than the ToDate.

current date - "DaysBeforeDepartureFrom" days &lt; current date -
"DaysBeforeDepartureTo", so DaysBeforeDepartureFrom &gt;
DaysBeforeDeparture</xs:documentation