New base schemas for export and import
Refactored BaseData.xsd into BaseDataImport.xsd and BaseDataExport.xsd .
In the Export the fields for the pricing differ to the fields for the
import, and the export has only a small subset of the fields available
for the import. In order to keep the schemas relevant to the respective
workflow it made sense to change the structure here.